Extract isDangerousHref to a shared helper to break import cycle
InlineLink imported isDangerousHref from Anchor, but Anchor transitively
imports InlineLink (via Tooltip, Popover, PortalRoot, Provider,
Translation and renderWithFormatting), tripping the no-circular check in
test:deps. Move the helper to shared/helpers and re-export it from Anchor
so the public API and existing tests are unchanged.
